Change Frontend repo for now to get travis to work
Travis was failing because the frontend repo had the proto
submodule path screwed up. At least, this was one of the
reasons for failure of Travis. Since I didn't have access
to Ritul's repo, I forked it to mine, fixed the submodule
referencing, and changed the submodule path in this repo
to my fork.