Fix the way Config works. Make testing nicer
This commit fixes how the config works as well as how the modules
get initialized. From now on one can test functions individually
using the 'run test' or 'debug test' tools in Visual Studio Code.
This is a part of an effort to modularize code so that testing
becomes easier and more sensible. It's a bit messy right now.