Fix critical bugs in Server
Recent changes in the server's architecture had introduced some severe
bugs, one of which didn't let the server listen on the HTTPS port if
unfulfilled orders existed in the database.
Another bug was in the handling of market depth stream. Since this
wasn't used last year, it didn't create problems then. But it was
revealed while testing this year that some severe bugs were there.