CancelOrder won't do ghost removals of StopLoss now
orderBook.CancelOrder() removes unfulfilled qty of the order from
the depth. It used to work fine, but then we decided to show
market & stoploss orders in market depth. So that lead this
code to remove cancelled stoploss orders that weren't activated
from the depth, but they were never really added to the depth.