Include example code to fix MySQL deadlocks
Issue #250: Random deadlocks

On lots of debugging, it was found that tx.Save(model) fires 3
queries in the transaction:

1. Update TransactionsTable Set X=foo, Y=bar Where userID=1 and
stockID = 1.
2. Select * from TransactionsTable where userID=1 and stockID = 1.
3. Insert into TransactionsTable Values (1, 1, foo, bar).

It tries to update first. Update will return "0 rows affected" in 2
cases:
a. When no row matches the where clause.
b. When the new values are same as old values of the matching row(s).

In this case, Gorm fires Select to check if it was case b. Select
returns some rows if they existed earlier.

If Select returns empty set, gorm Inserts a new row.

However, due to the way MySQL works, it seems that Insert and Updates
in the same transaction somehow cause a deadlock:

https://bugs.mysql.com/bug.php?id=52020
https://jira.mariadb.org/browse/MDEV-17073

I'm still not sure if this is the cause of the error. Checking MySQL
logs, I found that two transactions conflicted on same lock, which
shouldn't happen since we're using PRIMARY KEY index in each query,
and each query resolves to exactly one row, and MySQL docs say that
InnoDB will lock only that row's index record. Since each thread of
ours _does_ access different rows, I'm not sure why a common lock
even gets involved.

But life is such, and it does get involved, and I can't figure out
what the problem is.

So one solution is to not use tx.Save(...) and thus entirely avoid
Update-Insert issues. This commit shows that deadlock is avoided
by using tx.Where(...).Count(...) first to check if the row exists,
and if it does, it calls tx.Update(...), otherwise it calls
tx.Create(...)

This shouldn't really be required, and maybe we should file a bug
report on Gorm. It's further inefficient to use Update-Save-Insert
which uses 3 queries instead of using 2.

Signed-off-by: Parth Thakkar <pathakka@in.ibm.com>